refactor: move GameSession and GameException to shared
Moved the core domain logic machine GameSession and the relevant GameException into the shared module.
